+++ compiler-tests/src/test/data/dump/fir/private-provides-status-transformation/ProvidersAreMadePrivate.kt
@@ -1,0 +1,21 @@
+interface ExampleProviders {
+  @Provides
+  fun shouldBePrivate(): String = "hello"
+
+  @Provides
+  public fun shouldNotBePrivate1(): String = "hello"
+
+  @Provides
+  private fun shouldBePrivate2(): String = "hello"
+
+  companion object {
+    @Provides
+    fun shouldBePrivate(): String = "hello"
+
+    @Provides
+    public fun shouldNotBePrivate1(): String = "hello"
+
+    @Provides
+    internal fun shouldNotBePrivate2(): String = "hello"
+  }
+}